move container creation to node and enable cni
This moves container creation to the node service to have
a single service responsible for create and delete.  This
also switches to use CNI for networking.  By default, a loopback
device and a veth via the CNI bridge plugin will be created.
A custom Stellar IPAM service has also been created to allocate
IPs using the cluster wide subnets.

Signed-off-by: Evan Hazlett <ejhazlett@gmail.com>