Move context cancel into grpc-gateway goroutine
As the grpc-gateway http server is running in a goroutine, we need to
ensure that the context cancel is deferred until after the http server
stops running. I believe that at the moment, the http server starts in
the goroutine and then shortly after we call a deferred context cancel
at the end of the service's Start function call.