improve subnet allocation for nodes
This adds the ability to detect subnet changes in the configuration
as well as properly return errors when too small of subnet is
requested in the config (less than /22).  Since the network is an
additional network and nodes are routers we require that it be at least
a /21 to properly allow for a suitable number of network enabled
containers per host.

Signed-off-by: Evan Hazlett <ejhazlett@gmail.com>